# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FMAA Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ResourceCategory = typeof ResourceCategory[keyof typeof ResourceCategory];


export const ResourceCategory = {
  cv: 'cv',
  cover_letter: 'cover_letter',
  alumni_insight: 'alumni_insight',
  technical: 'technical',
  recruiting: 'recruiting',
} as const;

export type MembershipTier = typeof MembershipTier[keyof typeof MembershipTier];


export const MembershipTier = {
  standard: 'standard',
  premium: 'premium',
} as const;

export type MemberRole = typeof MemberRole[keyof typeof MemberRole];


export const MemberRole = {
  member: 'member',
  admin: 'admin',
} as const;

export interface Me {
  id: string;
  email: string;
  name: string;
  role: MemberRole;
  tier: MembershipTier;
  isPremium: boolean;
  isAdmin: boolean;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface Resource {
  id: number;
  title: string;
  slug: string;
  category: ResourceCategory;
  summary: string;
  content: string;
  tags?: string[];
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  coverImageUrl?: string | null;
  /** @nullable */
  readingMinutes?: number | null;
  isPremium: boolean;
  authorName: string;
  createdAt: string;
  updatedAt: string;
}

export interface ResourceInput {
  /** @minLength 1 */
  title: string;
  category: ResourceCategory;
  summary: string;
  content: string;
  tags?: string[];
  fileUrl?: string;
  coverImageUrl?: string;
  readingMinutes?: number;
  isPremium?: boolean;
}

export interface ResourceUpdate {
  title?: string;
  category?: ResourceCategory;
  summary?: string;
  content?: string;
  tags?: string[];
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  coverImageUrl?: string | null;
  /** @nullable */
  readingMinutes?: number | null;
  isPremium?: boolean;
}

export interface AlumniProfile {
  id: number;
  name: string;
  role: string;
  company: string;
  industry: string;
  gradYear: number;
  insight: string;
  /** @nullable */
  headshotUrl?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  location?: string | null;
  createdAt: string;
}

export interface AlumniInput {
  name: string;
  role: string;
  company: string;
  industry: string;
  gradYear: number;
  insight: string;
  headshotUrl?: string;
  linkedinUrl?: string;
  location?: string;
}

export interface AlumniUpdate {
  name?: string;
  role?: string;
  company?: string;
  industry?: string;
  gradYear?: number;
  insight?: string;
  /** @nullable */
  headshotUrl?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  location?: string | null;
}

export interface IndustryCount {
  industry: string;
  count: number;
}

export interface CategoryCount {
  category: ResourceCategory;
  count: number;
}

export interface DashboardSummary {
  totalResources: number;
  totalAlumni: number;
  categoryCounts: CategoryCount[];
  recentResources: Resource[];
  featuredAlumni: AlumniProfile[];
}

export interface Member {
  id: string;
  email: string;
  name: string;
  role: MemberRole;
  tier: MembershipTier;
  createdAt: string;
}

export interface MemberUpdate {
  role?: MemberRole;
  tier?: MembershipTier;
}

export interface ApprovedMember {
  email: string;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  addedBy?: string | null;
  createdAt: string;
}

export interface AddApprovedMembersInput {
  emails: string | string[];
  note?: string;
}

export interface AddApprovedMembersResult {
  added: number;
  emails: string[];
}

export type ListResourcesParams = {
category?: ResourceCategory;
q?: string;
};

export type ListAlumniParams = {
industry?: string;
};