# lib/api-zod/src/generated/types/memberUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FMAA Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { MemberRole } from './memberRole';
import type { MembershipTier } from './membershipTier';

export interface MemberUpdate {
  role?: MemberRole;
  tier?: MembershipTier;
}