# lib/api-zod/src/generated/types/resource.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FMAA Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { ResourceCategory } from './resourceCategory';

export interface Resource {
  id: number;
  title: string;
  slug: string;
  category: ResourceCategory;
  summary: string;
  content: string;
  tags?: string[];
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  coverImageUrl?: string | null;
  /** @nullable */
  readingMinutes?: number | null;
  isPremium: boolean;
  authorName: string;
  createdAt: Date;
  updatedAt: Date;
}