# lib/api-zod/src/generated/types/resourceCategory.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FMAA Portal API
 * OpenAPI spec version: 0.1.0
 */

export type ResourceCategory = typeof ResourceCategory[keyof typeof ResourceCategory];


export const ResourceCategory = {
  cv: 'cv',
  cover_letter: 'cover_letter',
  alumni_insight: 'alumni_insight',
  technical: 'technical',
  recruiting: 'recruiting',
} as const;