# lib/api-zod/src/generated/types/resourceInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FMAA Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { ResourceCategory } from './resourceCategory';

export interface ResourceInput {
  /** @minLength 1 */
  title: string;
  category: ResourceCategory;
  summary: string;
  content: string;
  tags?: string[];
  fileUrl?: string;
  coverImageUrl?: string;
  readingMinutes?: number;
  isPremium?: boolean;
}